refactor: using a better name for Entity
