+++ lib/index.ts
@@ -1,5 +1,5 @@
 export * from './shared';
 export * from './property';
 export * from './type-functions';
-export * from './entity';
+export * from './definition';
 export * from './module';
