+++ lib/shared.ts
@@ -1,7 +1,5 @@
 import { Schema } from 'mongoose';
 
-export type Definition = any;
-
 export const ObjectId = Schema.Types.ObjectId;
 
 export type ApiType =
