+++ tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "allowJs": true,
     "module": "commonjs",
+    "outDir": "./dist",
     "declaration": true,
     "noImplicitAny": false,
     "removeComments": true,
